[fix/#117] 홈 배너 클릭 삭제
