+++ app/src/main/java/com/poti/android/presentation/party/home/DummyHomeData.kt
@@ -1,0 +1,30 @@
+package com.poti.android.presentation.party.home
+
+import com.poti.android.domain.model.home.GroupItem
+
+val fakeMyGroupItems = listOf(
+    GroupItem(
+        postTitle = "2026 시즌 콘서트 후드",
+        artist = "아이유",
+        artistId = 0L,
+        postImage = "",
+        postCount = 3,
+        tag = "인기",
+    ),
+    GroupItem(
+        postTitle = "공식 응원봉 Ver.2",
+        artist = "아이유",
+        artistId = 0L,
+        postImage = "",
+        postCount = 12,
+        tag = "NEW",
+    ),
+    GroupItem(
+        postTitle = "월드투어 포토북",
+        artist = "아이유",
+        artistId = 0L,
+        postImage = "",
+        postCount = 7,
+        tag = "",
+    ),
+)
