+++ app/src/main/java/com/poti/android/presentation/party/home/navigation/HomeNavigation.kt
@@ -8,7 +8,6 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.poti.android.core.navigation.Route
 import com.poti.android.presentation.party.create.navigation.navigateToPartyCreate
-import com.poti.android.presentation.party.detail.navigation.navigateToPartyDetail
 import com.poti.android.presentation.party.goodsfilter.navigation.navigateToGoodsCategory
 import com.poti.android.presentation.party.goodsfilter.navigation.navigateToGoodsPartyList
 import com.poti.android.presentation.party.home.HomeRoute
@@ -26,7 +25,6 @@ fun NavGraphBuilder.homeNavGraph(
     composable<HomeRoute.Home> {
         HomeRoute(
             onNavigateToPartyCreate = navController::navigateToPartyCreate,
-            onNavigateToPartyDetail = navController::navigateToPartyDetail,
             onNavigateToGoodsPartyList = navController::navigateToGoodsPartyList,
             onNavigateToGoodsCategory = navController::navigateToGoodsCategory,
             modifier = Modifier.padding(paddingValues),
